Adds TerrainToolkit, much better RiverMaker.
RiverMaker scene works for generating a river in the scene. The
script is on the camera and there are a couple of settings that can
be tweaked. Reset the terrain by selecting it and using the Terrain
Toolkit's Create -> Perlin -> Generate Perlin Terrain.

Changes in game mode persist to edit mode; property of the Terrain
object.