Adds LoadRiverTest scene.
LoadRiverTest has a blank terrain and a new RiverManager script
attached to the camera. This script holds an array of JSON text
assets that are loaded when the scene starts and reconstituted into
River objects.

The RiverManager has public NextRiver, PrevRiver, and Redraw functions
for selecting a loaded River object and drawing it onto the terrain.
Once selected, the CurrentRiver property will expose the River object.